+++ CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning][semver].
 
 # Unreleased
 
-TMS-962-2: Register layout to frontpage components.
+- TMS-962-2: Register layout to frontpage components.
+- TMS-987:
+    - Change start_datetime & end_datetime fields to DateTimePicker & add functionalities for time
+    - Fix lunch-block & component not working if start_datetime was before the first lunch date
+
 # Released
+
 ## [1.0.0] - 2023-09-15
 
 - TMS-962: First stable version.