+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## Unreleased
 
+## [1.1.0] - 2023-11-07
+
 - TMS-987:
     - Change start_datetime & end_datetime fields to DateTimePicker & add functionalities for time
     - Fix lunch-block & component not working if start_datetime was before the first lunch date